witch back to using form_row() so that every field is rendered in a consistent way, unless you actually need to do something custom.
So those are your brave and valiant form rendering functions. Each function's first argument is something called 'view' - that's just the field - like genusForm.name. We call it a view, and you'll find out why later.

But check this out! Most of these functions also have an argument called variables. I know, that's a really generic-sounding argument. But it turns out that when it comes to kicking butt with form rendering, these variables are the key.