Field Options Versus Variables
Head back into the profiler. The subFamily field has a variable called placeholder set to 'Choose a Sub Family'. To see what this does, remove the disabled variable. Then, refresh. There it is! The placeholder is the option that appears at the top of the select element.

Why is this set to 'Choose a Sub Family'? Because that's what we passed as the placeholder option when we configured the field.
Back in the template, override the variable, set placeholder to Select a Subfamily.
So, which will win? The placeholder option, or the placeholder variable? Let's find out! Refresh!

It's 'Select a Subfamily': the variable wins.

I wanted to show you this because this touches on a really important thing. When you configure a field in your form class, each field has a set of options. These are not the same thing as the variables you can override in your template.

Nope: your form class holds field options, and your rendering functions have variables. Occasionally, a field has an option and a variable with the same name, like placeholder. But for the most part, these are two totally separate ideas: a field has a set of options, which mostly influence how the field should function, and a different set of variables, which help decide how the field will be rendered.