We've just added an X glyphicon to every field that has an error...
 12 lines  app/Resources/views/_formTheme.html.twig
{% block form_row -%}
    {% set showErrorIcon = (not compound or force_error|default(false)) and not valid %}
    <div class='form-group {% if (not compound or force_error|default(false)) and not valid %} has-error{% endif %}{{ showErrorIcon ? ' has-feedback' : '' }}'>
... lines 4 - 5
        {% if showErrorIcon %}
            <span class='glyphicon glyphicon-remove form-control-feedback' aria-hidden='true'></span>
        {% endif %}
... line 9
    </div>
{%- endblock form_row %}
And then found out that we really only want to add this to input fields.

How can we do that? I know that this is a text type. So maybe, instead of adding the icon to form_row, we could override the text_widget block and add it there. That would only affect text fields.

Go into form_div_layout.html.twig and look for text_widget. Woh! It's not here! That means Symfony must be using form_widget. That block does exist:

 372 lines  vendor/symfony/symfony/src/Symfony/Bridge/Twig/Resources/views/Form/form_div_layout.html.twig
... lines 1 - 2
{%- block form_widget -%}
    {% if compound %}
        {{- block('form_widget_compound') -}}
    {% else %}
        {{- block('form_widget_simple') -}}
    {% endif %}
{%- endblock form_widget -%}
... lines 10 - 372
Remember that compound variable I refused to explain before. Well, here it is again! We normally think of a field as just, well, a field: like a text box, or a select element. But sometimes, a field is actually a collection of sub-fields. An easy example is Symfony's DateType, which by default renders as 3 select elements for year, month and day. In that case, the DateType is said to be compound: it's just a wrapper for its three child fields.

In our form, all of our fields right now are simple: so, not compound. The block() function says:

Hey! Go render this other block called form_widget_simple.
After following all of this, it turns out that if we want to override the text widget, we need to override form_widget_simple:

 372 lines  vendor/symfony/symfony/src/Symfony/Bridge/Twig/Resources/views/Form/form_div_layout.html.twig
... lines 1 - 10
{%- block form_widget_simple -%}
    {%- set type = type|default('text') -%}
    <input type='{{ type }}' {{ block('widget_attributes') }} {% if value is not empty %}value='{{ value }}' {% endif %}/>
{%- endblock form_widget_simple -%}
... lines 15 - 372
In fact, all input fields - like the number field, search field or URL field - use this same block.

Ok, let's override it! But wait - check to see if it's in the Bootstrap template first:

 246 lines  vendor/symfony/symfony/src/Symfony/Bridge/Twig/Resources/views/Form/bootstrap_3_layout.html.twig
... lines 1 - 4
{% block form_widget_simple -%}
    {% if type is not defined or type not in ['file', 'hidden'] %}
        {%- set attr = attr|merge({class: (attr.class|default('') ~ ' form-control')|trim}) -%}
    {% endif %}
    {{- parent() -}}
{%- endblock form_widget_simple %}
... lines 11 - 246
It is! Copy that version, and paste it into _formTheme.html.twig.