The Craziness of Twig in Form Themes Now, check out this logic: form theme templates will have some of the craziest Twig code you'll ever see! In normal words, this says:
If type is not defined or file does not equal type, add a new form-control class.
To make this happen, it uses the attr variable that we were playing with before and merges in the new class, adding a space in case there was already a class.

Stealing Parent Blocks with use

Before we make any changes, go back and refresh. Woh! It doesn't work - that's surprising. The problem is this parent() call:

 246 lines  vendor/symfony/symfony/src/Symfony/Bridge/Twig/Resources/views/Form/bootstrap_3_layout.html.twig
... lines 1 - 4
{% block form_widget_simple -%}
... lines 6 - 8
    {{- parent() -}}
{%- endblock form_widget_simple %}
... lines 11 - 246
We understand that in normal Twig templates, you can override parent blocks and use the parent() function. But check this out: our template does not extend any Twig template... and we don't want it to! For reasons that honestly aren't very important, a form theme template should never extend anything.

But wait, then, how did this code work in the Bootstrap template? Go look: at the top, it has a use for form_div_layout.html.twig:

 246 lines  vendor/symfony/symfony/src/Symfony/Bridge/Twig/Resources/views/Form/bootstrap_3_layout.html.twig
{% use form_div_layout.html.twig %}
... lines 2 - 246
The use says:

I don't actually want to extend this other template. But, please allow me to call the parent() function as if I were extending it.