The Error Icon in the Widget Finally, we can move the icon to this block. First, keep that showErrorIcon variable: we need that to add the has-feedback class. But copy it and move it down into form_widget_simple, inside the if statement... because, it turns out, we probably also don't want to show the error icon if this is a file upload field.
Above, set showErrorIcon to false by default.

Finally, copy the span icon code, remove it, and paste it right after the parent call, to put this after the widget.

That should do it! Resubmit the form! Got it! One fancy error icon on the name text field, and zero fancy error icons on the select field.

In a nutshell, form theming means:

(A) finding the right block to override and then;
(B) leveraging your variables to do cool stuff.
Next, we'll add a missing feature to Symfony: field help text.