Look closer at the Bootstrap help feature: to be nice to screen readers, we should add an aria-describedby attribute to the field that points to an id that we add to the help span:
<label class='sr-only' for='inputHelpBlock'>Input with help text</label>
<input type='text' id='inputHelpBlock' class='form-control' aria-describedby='helpBlock'>
...
<span id='helpBlock' class='help-block'>
    A block of help text that breaks onto a new line and may extend beyond one line.
</span>

That way, when a screen reader focuses on the text box, it will read the help text, which is pretty rad. It also turns out that pulling this off is a cool challenge!

Let's start with a plan: when the form_widget() function is called inside form_row:

 26 lines  app/Resources/views/_formTheme.html.twig
... lines 1 - 2
{% block form_row -%}
... line 4
    <div class='form-group {% if (not compound or force_error|default(false)) and not valid %} has-error{% endif %}{{ showErrorIcon ? ' has-feedback' : '' }}'>
... line 6
        {{- form_widget(form) -}}
... lines 8 - 11
    </div>
{%- endblock form_row %}
... lines 14 - 26
we want the attr variable to have a new key called aria-describedby. We've seen magic like this before: in the Bootstrap layout, the form_widget_simple block modifies the attr variable before calling the parent block:

 246 lines  vendor/symfony/symfony/src/Symfony/Bridge/Twig/Resources/views/Form/bootstrap_3_layout.html.twig
... lines 1 - 4
{% block form_widget_simple -%}
    {% if type is not defined or type not in ['file', 'hidden'] %}
        {%- set attr = attr|merge({class: (attr.class|default('') ~ ' form-control')|trim}) -%}
    {% endif %}
    {{- parent() -}}
{%- endblock form_widget_simple %}

That's what we want to do!

Modifying the attr variable

Back in our block, before form_widget(), add another if help|default. Inside, set attr = attr|merge() with an array argument. The core merge filter will array_merge() the argument back into the attr variable. Add aria-describedby set to... well, nothing yet.

First, we need to give our help span an id. Do that: set it to help-block- then print the id variable to make sure this is unique.