The id variable will become the id attribute on the field itself. Now set the aria-describedby to help-block-, a ~ then id.
The ~ is Twig's rarely-used concatenation operator, so, it's like . in PHP.

Ok! Now that attr is changed before we call form_widget, it'll hopefully render on that widget. Time to give it a try. Refresh!

Ok, go dig into the source to see if the attribute is there. Umm... it's not! There is not any aria-describedby. This tutorial is a LIE!