No no, it's cool. It turns out that there's a very subtle, but important detail that I'm neglecting. Let me show you: click to open the parent form_div_layout.html.twig template. We're letting Symfony guess this, but the speciesCount is a NumberType, meaning it'll render as an <input type='number' /> field.
Inside the layout file, find the number_widget block that renders this:

 372 lines  vendor/symfony/symfony/src/Symfony/Bridge/Twig/Resources/views/Form/form_div_layout.html.twig
... lines 1 - 133
{%- block number_widget -%}
    {# type=number doesn't work with floats #}
    {%- set type = type|default('text') -%}
    {{ block('form_widget_simple') }}
{%- endblock number_widget -%}
... lines 139 - 372

Ok, check it out: it sets a type variable, and then calls the form_widget_simple block. Then, when form_widget_simple executes, type is set to number:

 372 lines  vendor/symfony/symfony/src/Symfony/Bridge/Twig/Resources/views/Form/form_div_layout.html.twig
... lines 1 - 10
{%- block form_widget_simple -%}
    {%- set type = type|default('text') -%}
    <input type='{{ type }}' {{ block('widget_attributes') }} {% if value is not empty %}value='{{ value }}' {% endif %}/>
{%- endblock form_widget_simple -%}
... lines 15 - 372

So, why does that work, but not our code? Well look at that code again: it sets a variable and then calls the block function. When you call block(), all variables flow through to that block.

But now check out our code: we set a variable, but then we don't execute a block! We call form_widget(). Hey! that's a form rendering function - the same kind that we use inside our normal templates. In this case, the variables do not magically flow through. But that's ok! We already know how to pass variables into form_widget(). Add a second argument, and pass attr set to attr.
Let's refresh! Inspect the field, well, not any field - inspect the isPublished field. This time, we got it!