Overriding Field Behavior
Here's the cool thing about these classes: they have all the same functions as a normal form class, like buildForm() or configureOptions(). The difference is that whatever modifications we make to this class will literally be applied to every field in the system.

For example, go back to the 'Code'->'Generate' menu, click 'Override Methods', then select buildView().

When we're done setting things up, whenever any field is transformed into a FormView object, this method will be called and we will be able to add variables to anything!

Try it: add {dollar_sign}view - which represents whatever one field is being setup - {dollar_sign}view->vars['help'] set to TURTLES!

That's a ridiculous, and yet, fully-functional type extension.