So, we're done! Wait... except for this random field at the bottom of my form. What the heck? That's my year field!
See that form_end() at the bottom of our form?

Remember how it renders any field that we forgot to render? Well, now it thinks that we forgot to render the year field. The nerve!

So, could we just tell it that the field was actually rendered? Yep, and the code is both simple and strange. Use a rare do tag from Twig and say genusForm.firstDiscoveredAt.year.setRendered().

Well, every field is a FormView object. And if you open that class, it has a setRendered() method!
163 lines  vendor/symfony/symfony/src/Symfony/Component/Form/FormView.php
... lines 1 - 18
class FormView implements \ArrayAccess, \IteratorAggregate, \Countable
{
... lines 21 - 86
    /**
     * Marks the view as rendered.
     *
     * @return FormView The view object
     */
    public function setRendered()
    {
        {dollar_sign}this->rendered = true;
        return {dollar_sign}this;
    }
... lines 98 - 161
}

And by calling it, we're saying:

Yo, we rendered this already. So, you know, don't try to render it again.
Refresh now! Now that extra field is gone.